Remove components that were moved to shared and react-components
