+++ src/app/containers/AutoReplyContainer.js
@@ -1,7 +1,25 @@
 import React from 'react';
+import { ObserverSections, AutoReplySection, RelatedSettingsSection } from 'react-components';
+import { c } from 'ttag';
 
 const AutoReplyContainer = () => {
-    return <h1>AutoReplySection</h1>;
+    return (
+        <ObserverSections>
+            <AutoReplySection id="auto-reply" />
+            <RelatedSettingsSection
+                id="related-settings"
+                list={[
+                    {
+                        icon: 'filter',
+                        text: c('Info')
+                            .t`Go to Filter Settings if you want to create and manage auto-replies by email address.`,
+                        link: c('Link').t`Filter Settings`,
+                        to: '/settings/filters'
+                    }
+                ]}
+            />
+        </ObserverSections>
+    );
 };
 
 export default AutoReplyContainer;
