+++ src/app/content/AuthenticatedAppRoutes.js
@@ -18,7 +18,7 @@ import KeysContainer from '../containers/KeysContainer';
 import GeneralContainer from '../containers/GeneralContainer';
 import FiltersContainer from '../containers/FiltersContainer';
 import LabelsContainer from '../containers/Labels/Container';
-import AutoReplyContainer from '../containers/AutoReply/Container';
+import AutoReplyContainer from '../containers/AutoReplyContainer';
 
 const NotFoundContainer = () => <h1>Not found</h1>;
